Added SKIL daemon client dependency. Moved endpoints Main class and created skil client main class
