+++ endpoint_clients/src/main/java/ai/skymind/skil/examples/endpoints/Main.java
@@ -1,4 +1,4 @@
-package ai.skymind.skil.examples;
+package ai.skymind.skil.examples.endpoints;
 
 import ai.skymind.skil.examples.endpoints.*;
 import org.json.JSONArray;
