+++ endpoint_clients/src/main/java/ai/skymind/skil/examples/skil_client/Main.java
@@ -1,0 +1,7 @@
+package ai.skymind.skil.examples.skil_client;
+
+public class Main {
+    public static void main(String[] args) {
+
+    }
+}
